tests: do a deep comparison of options in _compare_rrsets
The current code compares RRsets by comparing the output of the ->string
function, which has remained stable.

That is, until Net::DNS 1.36, where the string output of e.g. CLIENT-SUBNET is
the JSON representation of a hash with its components as members (family,
scope, etc.), as constructed by _format_option internally. With hashes in Perl
being unordered, this creates non-deterministic output.

Code in the SVN trunk suggests there is now a sort, to sort the hash by key.
However, this code is not yet released, and 1.36 can be found in the wild,
including in distributions like Debian bookworm.

Therefore, add code to do a deep-level comparison instead, ensuring that the
options of each RRset match each other.

Backwards-compatible, tested with 1.05, and 1.32 - 1.36.