Fix test compatibility issue with Net::DNS-1.33+
With the release of Net::DNS 1.33, a "bug" was fixed, and the
resolver recurse attribute no longer sets the RD bit in the query
packet for us.  We've switched tactics here to set it directly
instead.

Net::DNS 1.33 is the version shipped in ubuntu 22.04, which is
what Github CI is currently using for "ubuntu-latest", and this
should thus hopefully fix the test failure we're seeing there.