Sonar: Rely on props file, explicit build-wrapper
It seems like the workflow is picking up sonar-project.properties
anyways, so sync the current changes back over to there, remove
the copies in the workflow file.

Also, I'm starting to think the github action for analysis won't
run the build wrapper for me (which kinda makes sense!), so let's
see if we can run it explicitly, and install all the build/test
deps like our c-cpp workflow does as well.