Require sendmmsg support
Linux and all the major BSDs have supported recvmmsg() +
sendmmsg() interfaces for a while now, and the timing seems like
it might be about right to drop support for the legacy
sendmsg()/recvmsg() alternative, given it's 2023 and we're still
probably months out from 4.0 being a real thing.

The UDP "mainloop" code is pretty complex and perf-critical, so
one of the benefits here is getting rid of the somewhat redundant
definition of for the single-packet variant.  It has been a pain
to keep them both at parity over time.

The obvious downside, of course, is there might be some platform
which is still relevant and useful for a gdnsd server which
doesn't support these interfaces.  I figure this is worth a shot,
we can always revert later if there's enough complaint.