tests: fix AXFR-related Net::DNS 1.39 breakage
Net::DNS 1.39 made axfr() ignore the "persistent_tcp" setting, which
means that the socket is closed right after the AXFR, which in turn
increases tcp_close_c.

This was previously not the case, and hence the 003complex/009broken.t
tests right after the AXFR (7 onwards) fail with:
  # Stats check: Stats check failed: tcp_close_c mismatch (hard-fail), wanted 0, got 1

Add tcp_close_c to the AXFR test's stats_inc, which resolves this issue
for 1.39+.

To ensure backwards-compatibility with older versions, also add some
code to turn off persistent_tcp() before the axfr() call, and restore it
to its previous state right after (currently always 1).

Fixes #236