Fix process management issues
This is a revert of a small part of commit e6760666 "Robustness
improvements for replacement launching", which was included in the
v3.8.1 release.

The bulk of that commit seems fine, but the process management
changes (removal of waitpid() src/css.c, ignoring SIGCHLD in the
main process, removing the libev child reaper) were misguided, and
caused two issues:

1. "gdnsd daemonize" didn't work right at all.  We don't commonly
   use this on Linux/systemd systems anymore and the testsuite
   doesn't exercise it either, so this went unnoticed until a
   FreeBSD user reported it!

2. It caused a (probably very unlikely in practice!) race
   condition during "replace" under systemd, which could result in
   the new replacement daemon being considered an "alien" main pid
   by systemd because it hadn't been re-parented to init before it
   notified systemd of readiness and mainpid change.  On
   relatively recent versions of systemd, this doesn't seem to
   cause any functional issue (merely a spurious error message),
   but on older systemd installations it can prevent correct
   operation, IIRC.

This commit restores these matters to their previous state:
SIGCHLD is no longer ignored, the libev child repear is used, and
we explicitly waitpid() for the middle child process during
"replace" to avoid the above-mentioned alien main pid race.

Fixes #238
See also #243