Ignore SIGCHLD during kill_registered_children()
At some point in the recent history refactoring the main eventloop
and extmon child process stuff, I managed to create a situation
where we always wait 15 seconds near the end of a clean daemon
shutdown if extmon is in use.  This is because the killing of the
extmon_helper child process happens after the libev loop has
exited, and therefore nothing reaps its status to allow our kill()
probes to return zero.

This problem was further masked in 3.8.1 by the other process
management changes which included ignoring SIGCHLD for the life of
the daemon, and which have since been reverted.

Since we don't particularly care about its final disposition
anyways (extmon_helper is the only case, and does log its own
progress anyways), the simplest solution is to ignore SIGCHLD here
on the way down.  You could argue for a complex, staged waitpid
setup with a timeout, but this code is destined to be deprecated
and later die eventually anyways, and this is good enough for now.