qsort pedantry
People will be looking for such cases after seeing
https://www.qualys.com/2024/01/30/qsort.txt , so I may as well
refactor this case to be more idiomatic.

Note there was no actual issue here, as (a) the natural x->mask
values are validated by nets_parse() to be in the range 0-128 and
(b) the normalizer can also set a magic value for deletions, but
this was also just under half the bits of an unsigned (0xFFFF).

I've changed the magic value to 0xFF to make things cleaner (now
all values fit in 8 bits, instead of having the one outlier as a
16-bit value), and refactored the qsort callback to both assert
the expected mask value range and use the more-idiomatic
integer comparison method of: (a > b) - (a < b).